rename ensurePemKeyPair -> certificateManager
